refactor: separate Medical from SingleCardLayout
